Fix the pageqlock recusion on sun4m.
In uvm_pageout() we called pmap_page_protect(VM_PROT_NONE) with the
pageqlock held, when pageqlock was a mutex this caused recursion panics.
The problem was that freeing the pv_entries called uvm_km_free, which
needed the pageqlock itself. Fix this by changing the homegrown pool
allocator to just use the normal pool functions, then change the
cacheability as necessary.

That I know of, this leaves m68k still to be fixed, maybe arm bootstrap
(but I do not think so).

Tested by NicM@ on a 4m machine. ok miod@