make mpii properly detach devices, which helps a lot if they have commands in flight. to relevant changes are:
- call the activate(DVACT_DEACTIVATE) function against all the luns
on the target that is going away as soon as possible.
- issue the target reset BEFORE detaching the children devices.
this is needed now tha the midlayer will sleep until all outstanding
commands on a device come back from the adapter before calling the
child devices attach routine.

tested on straight disks and on disks in enclosures.

ok and moral support from krw@