fix bug for real. Turns out some libraries (xcb...) have NEEDED that go to standard ldconfig search dirs...
Of course, ld complains because it doesn't have them in its standard
search dir. BUT there's an option -rpath-link designed just for that
(look up needed objects during compile and trust the end result to
find them).

So use that for ldconfig directories... so that xcb can link without
needing to go have a look at the NEEDED list, but without encoding
standard rpaths in the resulting binary...

Also, put all the junk that needs -Wl together so we can add one single
-Wl,-rpath,foo,-rpath-link,foo2  thingy...