Add support for Microsoft XBox 360 controller as a uhid. It doesn't use the standard interface class and doesn't have a report descriptor, so use a manually created one.
The controller shows up as multiple devices with different interface
numbers, so only match the first one as a uhid.  The controller shows up
as both a uhid and a ugen.

Rename ugraphire_rdesc.h to uhid_rdesc.h and add the manually created
Xbox 360 report descriptor to the file.

Add ifndef SMALL_KERNEL around the related blocks for the XBox 360
controller and Wacom Graphire, so this should actually shrink the
ramdisks.

Manual report descriptor taken from a 2005 FreeBSD commit.  Much help
from mpi@ for getting this suitable for commit to OpenBSD.

OK mpi@