Do not add -lcompat in configure.
libcompat isn't needed; configure checks for the presence of cuserid()
and ftime() and the build copes well with the absence of either.
ftime() would only ever be used if gettimeofday() wasn't available,
and cuserid() is replaced by getlogin() but that code branch is not
reached anyway.

ok sthen@ jca@ deraadt@