if it looks like a variable, expands like a variable and resolves like a variable, then it need not be a variable.
add a way out the variable expansion loop when the returned value from the
values hash looks like a variable.

this resolves an issue where using --define-variable=libdir=${libdir} would
create and infinite loop of looking up the value of libdir. such as triggered
by the x11/gnome/empathy configure script.

tested in a bulk build as well as builds of base and xenocara
ok aja@