If the PDC_COPROC call fails, and we are running on a PA 1.1 system, assume the usual 1.1 coprocessor mask, instead of using zero and having every userland process dying with signal 4.
The circumstances under which this call fails are unclear and could be
tied to specific PROM version (I have tried overzealous stack alignment and
other tricks, to no avail). Interestingly enough, the 715/75 system which
hits this problem, only triggers it when booting from disk, and never when
booting from network.

This diff is an ugly bandaid until the problem is better understood. Or maybe
it is not worth investigating, seeing that Linux hardcodes the coprocessor
mask and never issues PDC_COPROC calls; I wonder what HP/UX does.

In my tree for over 18 months; "Looks like a nice hack" deraadt@
715/75 system provided by Sebastiaan Indesteege, thanks!