Add the userland facing drm header for nouveau.
This is required by the generic loader code in newer versions of Mesa
even when we aren't building any nouveau drivers.

The way these headers are used on Linux is with a kernel version
and a libdrm version which are largely identical with the *_drm.h
headers always installed by libdrm when libdrm is built with autoconf.

In OpenBSD we have a single version of the _drm.h headers that
are shared by the kernel and userspace and have our own simple
Makefile for the libdrm build.